+++ src/app/app.component.html
@@ -5,7 +5,6 @@
       <a mat-list-item routerLink="/experiments">Experiments</a>
       <a mat-list-item routerLink="/media">Media</a>
       <a mat-list-item routerLink="/subjects">Subjects</a>
-      <a mat-list-item routerLink="/sessions">Sessions</a>
     </mat-nav-list>
   </mat-sidenav>
   <mat-sidenav-content class="nav-content">
