+++ src/app/app.module.ts
@@ -19,8 +19,7 @@ import { PapaParseModule } from 'ngx-papaparse';
 
 import { AppComponent } from './app.component';
 import { TimeSeriesComponent } from './time-series/time-series.component';
-import { ChartService } from './shared/chart.service';
-import { HeadViewComponent } from './head-view/head-view.component';
+import { ChartService } from './shared/services/chart.service';
 import { HeadsetInfoComponent } from './headset-info/headset-info.component';
 import { RecorderComponent } from './recorder/recorder.component';
 import { ExperimentFormComponent } from './experiment-form/experiment-form.component';
@@ -32,15 +31,12 @@ import { ExperimentHubComponent } from './experiment-hub/experiment-hub.componen
 import { DataCollectionComponent } from './data-collection/data-collection.component';
 import { SubjectFormComponent } from './subject-form/subject-form.component';
 import { SubjectHubComponent } from './subject-hub/subject-hub.component';
-import { SessionsComponent } from './sessions/sessions.component';
-import { SessionFormComponent } from './session-form/session-form.component';
 
 const appRoutes: Routes = [
   { path: 'home', component: HomeComponent },
   { path: 'experiments', component: ExperimentHubComponent },
   { path: 'media', component: DataCollectionComponent },
   { path: 'subjects', component: SubjectHubComponent },
-  { path: 'sessions', component: SessionsComponent },
   { path: '**', component: HomeComponent },
 ];
 
@@ -48,7 +44,6 @@ const appRoutes: Routes = [
   declarations: [
     AppComponent,
     TimeSeriesComponent,
-    HeadViewComponent,
     HeadsetInfoComponent,
     RecorderComponent,
     ExperimentFormComponent,
@@ -59,8 +54,6 @@ const appRoutes: Routes = [
     DataCollectionComponent,
     SubjectFormComponent,
     SubjectHubComponent,
-    SessionsComponent,
-    SessionFormComponent
   ],
   imports: [
     HttpClientModule,
