--- src/app/head-view/head-view.component.css
@@ -1,5 +1,0 @@
-mat-spinner {
-  position: absolute;
-  top: 160px;
-  left: 120px;
-}
