--- src/app/head-view/head-view.component.html
@@ -1,1 +1,0 @@
-<mat-spinner *ngIf="!modelLoaded"></mat-spinner>
