--- src/app/session-form/session-form.component.css
@@ -1,20 +1,0 @@
-.session-form {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .session-form > * {
-    width: 100%;
-  }
-  
-  .session-form form {
-    margin-bottom: 20px;
-  }
-  
-  .session-form form > * {
-    margin: 5px 0;
-  }
-  
-  .session-form .mat-radio-button {
-    margin: 0 5px;
-  }
