--- src/app/sessions/sessions.component.html
@@ -1,27 +1,0 @@
-<section>
-  <mat-card>
-    <mat-card-title>Sessions Hub</mat-card-title>
-    <mat-card-content>
-      <button (click)="newSession()" mat-raised-button>New Session</button>
-      <mat-list>
-        <mat-list-item *ngFor="let session of eService.sessions">
-          <span>
-            {{session.id}}
-            <button style="margin:0 20px 0 20px" (click)="editSession(session)" mat-button name="edit">
-              <mat-icon>edit</mat-icon>
-            </button>
-            <button (click)="deleteSession(session)" mat-button name="delete">
-              <mat-icon>delete</mat-icon>
-            </button>
-          </span>
-        </mat-list-item>
-      </mat-list>
-    </mat-card-content>
-  </mat-card>
-  <mat-card *ngIf="selectedSession">
-      <mat-card-title>Session Form</mat-card-title>
-      <mat-card-content>
-        <app-session-form [session]="selectedSession"></app-session-form>
-      </mat-card-content>
-    </mat-card>
-</section>
