+++ src/app/shared/classes/experiment.ts
