+++ src/app/shared/classes/media-description.ts
