+++ src/app/shared/classes/session.ts
