+++ src/app/shared/services/eeg-stream.service.spec.ts
