+++ src/app/shared/services/eeg-stream.service.ts
