+++ src/app/shared/services/file-player.service.ts
