+++ src/app/shared/services/subject.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Subject } from './subject';
+import { Subject } from '../classes/subject';
 import gql from 'graphql-tag';
 import { Observable, of, empty } from 'rxjs';
 import { Apollo, QueryRef } from 'apollo-angular';
